term: Move more mouse functions to mouse.c
