+++ src/nvim/ex_cmds.c
@@ -47,6 +47,7 @@
 #include "nvim/garray.h"
 #include "nvim/memory.h"
 #include "nvim/move.h"
+#include "nvim/mouse.h"
 #include "nvim/normal.h"
 #include "nvim/ops.h"
 #include "nvim/option.h"
