+++ src/nvim/ex_getln.c
@@ -48,6 +48,7 @@
 #include "nvim/keymap.h"
 #include "nvim/garray.h"
 #include "nvim/move.h"
+#include "nvim/mouse.h"
 #include "nvim/ops.h"
 #include "nvim/option.h"
 #include "nvim/os_unix.h"
