+++ src/nvim/search.c
@@ -39,6 +39,7 @@
 #include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/move.h"
+#include "nvim/mouse.h"
 #include "nvim/normal.h"
 #include "nvim/option.h"
 #include "nvim/path.h"
